src/main/java/com/hsbc/incident/config/JwtAuthenticationFilter.java
package com.hsbc.incident.config;

import com.hsbc.incident.domain.model.entity.User;
import com.hsbc.incident.domain.service.UserService;
import com.hsbc.incident.shared.constant.CommonConstant;
import java.io.IOException;
import java.util.UUID;
import javax.servlet.FilterChain;
import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import lombok.RequiredArgsConstructor;
import org.slf4j.MDC;
import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.context.SecurityContextHolder;
import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
import org.springframework.stereotype.Component;
import org.springframework.util.StringUtils;
import org.springframework.web.filter.OncePerRequestFilter;

@Component
@RequiredArgsConstructor
public class JwtAuthenticationFilter extends OncePerRequestFilter {

    private final UserService userService;

    public static final String BEARER_PREFIX = "Bearer ";
    public static final String AUTHORIZATION_HEADER = "Authorization";

    @Override
    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
        FilterChain filterChain) throws ServletException, IOException {

        MDC.put(CommonConstant.MDC_TRACE_ID, UUID.randomUUID().toString());

        String header = request.getHeader(AUTHORIZATION_HEADER);
        if (StringUtils.hasText(header) && header.startsWith(BEARER_PREFIX)) {
            String tokenStr = header.substring(BEARER_PREFIX.length());

            Long userId = userService.validateToken(tokenStr);
            request.setAttribute(CommonConstant.CONTEXT_LOGIN_USER_ID, userId);

            User user = userService.getUserById(userId);

            UsernamePasswordAuthenticationToken usernamePasswordAuthenticationToken =
                new UsernamePasswordAuthenticationToken(userId, tokenStr,
                    user.getPermissions().stream()
                        .map(Enum::name)
                        .map(SimpleGrantedAuthority::new)
                        .toList());
            usernamePasswordAuthenticationToken
                .setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
            SecurityContextHolder.getContext()
                .setAuthentication(usernamePasswordAuthenticationToken);
        }
        filterChain.doFilter(request, response);
    }
}
